requirements.txt
requests
ffmpeg-python
pillow
matplotlib